Create 1.0-batchprocessing_v2.0-minsizemod.ijm
As shown on lines 65 and 66, the run("Analyze particles", "...") command was modified to assess the effect of size parameters on the SS image segmentation macro (presented in lab meeting on 06/08/20). Three different size parameters were tested (min size = 0.25, 0.40, or 0.45) and each size parameter with a minimum size below 0.50 showed over-outlining of non-cell "objects" in the image background. Still troubleshooting why the over-outlining occurred.